ChirpFunction/requirements.txt
functions-framework==3.*
google-api-core==2.19.0
google-cloud-logging==3.10.0
google-cloud-speech==2.26.0
